+++ src/c64/VIC_II/mos656x.cpp
@@ -264,8 +264,8 @@ void MOS656X::event()
     if (cycles)
     {
         // Update x raster
-        rasterClk += cycles;
-        lineCycle += cycles;
+        rasterClk += (int)cycles;
+        lineCycle += (int)cycles;
         lineCycle %= cyclesPerLine;
 
         delay = (this->*clock)();
@@ -273,7 +273,7 @@ void MOS656X::event()
     else
         delay = 1;
 
-    eventScheduler.schedule(*this, delay - eventScheduler.phase(), EVENT_CLOCK_PHI1);
+    eventScheduler.schedule(*this, (unsigned int)(delay - eventScheduler.phase()), EVENT_CLOCK_PHI1);
 }
 
 event_clock_t MOS656X::clockPAL()
